:recycle: refactor(components,messages): refactor pagination, add i18n messages
