+++ src/components/common/pagination/index.ts
@@ -1,1 +1,1 @@
-export { default } from './pagination';
+export { default } from './components/pagination';
